new web ui code refactor
