+++ web-ui/src/shared/components/AmbientOrbs.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+export default function AmbientOrbs() {
+  return (
+    <>
+      <div className="orb orb-1" />
+      <div className="orb orb-2" />
+      <div className="orb orb-3" />
+    </>
+  );
+}
